fix: jwt 라이브러리 설치, 코드 분리
