+++ be/algo-with-me-api/src/auth/auth.module.ts
@@ -1,10 +1,11 @@
 import { Module } from '@nestjs/common';
 
 import { AuthController } from './auth.controller';
+import { AuthService } from './auth.service';
 import { GithubStrategy } from './auth.strategy';
 
 @Module({
   controllers: [AuthController],
-  providers: [GithubStrategy],
+  providers: [GithubStrategy, AuthService],
 })
 export class AuthModule {}
